pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Проект: Менеджер задач (Python)"
readme = "README.md"
requires-python = "~=3.12.0"
dependencies = [
    "dj-database-url>=3.1.2",
    "django>=6.0.2",
    "django-bootstrap5>=26.2",
    "django-filter>=25.2",
    "gunicorn>=25.1.0",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.1",
    "rollbar>=1.3.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.11",
]

[tool.ruff]
output-format = "concise"
line-length = 80
extend-exclude = ["__init__.py", "manage.py", "*/migrations/"]

[tool.ruff.lint]
extend-select = [
    "E", # pycodestyle Errors
    "W", # pycodestyle Warnings
    "C90", # mccabe
    "I", # isort
    "D", # pydocstyle
    "B", # flake8-bugbear
    "COM", # flake8-commas
    "Q", # flake8-quotes
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "PL", # Pylint
    "N", # pep8-naming
    "AIR", # Airflow
    "FAST", # FastAPI
    "DJ" # flake8-django
]
ignore = ["COM812"] # Конфликт с ruff format
preview = true

[tool.ruff.lint.per-file-ignores]
"tests.py" = ["RUF012"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.format]
quote-style = "single"
